+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+### 4.3.0(67)
+
+Services
+
+- `KeychainServiceConfiguration` is added to allow modification of queries in `KeychainService`
+- `KeychainStorage` no longer saves default value to Keychain, until it's retrieved for the first time
+- `class` subscripts in `KeychainService` are replaced with instance ones
+
 ### [4.2.0(66)](https://github.com/VakhoKontridze/VCore/releases/tag/4.2.0) — *2022 10 29*
 
 General